feat: Dao interface 제거
